nutrition_insights_backup_2025-08-22/phase3/requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
plotly>=5.20.0
pyyaml>=6.0
requests>=2.31.0
numpy>=1.24.0